Split out secure variables state store impl
